extra: improve how "if (foo & ~0xF)" is handled
In the previous code we said that if you had "if (foo & 0xf0) {" then we
knew that foo was a minimum of 0x10.  Now we're saying that if you have
"if (foo & ~0xff) " then we know that foo is a maximum of 0xff.

Looking at it now, I see that this could be tightened further still...  :/
There is a lot of work to do.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>