db/fixup_kernel.sh: work around for gfs2_ea_find()
It would be better if this wasn't needed, of course.  What happens is this:

We pass in "el".

        ef.ef_el = el;
        memset(el, 0, sizeof(struct gfs2_ea_location));
        error = ea_foreach(ip, ea_find_i, &ef);

Smatch sees that we memset "el".  But ea_foreach() treats "ef" as a void
pointer to data.  Smatch marks it as an untracked parameter.  But smatch
doesn't make the jump from ef.ef_el back to el.  So the caller says that
everything in el is set to zero.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>