db/fixup_kernel.sh: work around for __to_fd()
The __to_fd() function looks like this:

static inline struct fd __to_fd(unsigned long v)
{
        return (struct fd){(struct file *)(v & ~3),v & 3};
}

We should be recording it in the database as the place where fd->file and
fd->flags are set.  But we don't because that's hella complicated.  So
instead the database records a couple other places and determines that
fd->file is always NULL.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>