slist: Silence a smatch warning
Smatch complains that we previously assumed that one_iter.sm could be NULL.

The truth that if either one is NULL then the other is NULL so I could
just remove the second half of the check.  But I guess I want it in for
documentation to show that both are NULL when we exit, otherwise it looks
unbalanced.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>