let "compile" not crash on bools
"compile" & "compile-i386" have never been adapted to use
bits_to_bytes() and crash when processing a bool because
it doesn't find a suffix for 1-bit instructions ('b', 'w', 'l', 'q').

Sprinkling a dozen "bit_to_bytes()" and adjusting a few shifts
would fix it but letting bits_in_bool be equal to 8 is much simpler.

Note: with this patch, "compile" crashes 'only' 25 times when used
on the testsuite files, instead of 27 without it.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>
CC: Jeff Layton <jlayton@redhat.com>
Signed-off-by: Christopher Li <sparse@chrisli.org>