simplify '(x op x)' to '0', '1' or 'x'
Most binops can be simplified when their two operands are
identical. For example '(x ^ x)' can be simplified into '0'.

The cases '(x / x)' and '(x % x)' are not simplified since this
is correct only when x is not zero.
The cases '(x || x)' and '(x && x)' are not simplified because
it's only correct when the operands are booleans/have already
been compared against zero and the linearization don't enforce that.

This patch add the code for these simplifications as well as
their corresponding test cases.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>
Signed-off-by: Christopher Li <sparse@chrisli.org>