recursive phi_defines cannot happen
The function phi_defines() does the liveness tracking of phi-pseudos
(pseudos in OP_PHIs args). While doing this, the function is
recursively called when the pseudo would be defined by another phi-node;
but this condition is impossible because all phi-pseudos are defined
by OP_PHISOURCE instructions.

This patch remove the recursive call and its associated test.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>
Signed-off-by: Christopher Li <sparse@chrisli.org>