simplify '(x || x)' and '(x && x)'
The operators '||' and '&&' being idempotent, the expressions
'(x || x)' and '(x && x)' can be simplified to a test against zero.

Note: they could even be replaced by 'x' itself but only if 'x' is
already a boolean expression/has already been tested against zero.
If it is the case, the redundant test this will be optimized away
in further steps.

For example, test-linearize on the following code:
	int ior(int a) { return a || a; }
emitted the following instructions:
	or-bool.32  %r3 <- %arg1, %arg1
after the patch, it now emits:
	setne.32    %r3 <- %arg1, $0
which is easier to combine with others simplifications.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>
Signed-off-by: Christopher Li <sparse@chrisli.org>