db: handle function pointers better
The issue here is say you have a function merge_estate(), let's say, that's
called from several places.  Or another example is if there are two
functions with the same name.  Then those can get interleaved and the
solution is to sort them by filename.  It's potentially not a 100% solution
but it's 99%.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>