function_hooks, db: use INTERNAL as the marker for a new function return
The problem here is that you can have cases where two functions have the
same return_id so a new function is not detected.  Say you have a function
hook so really it's just the same function called from two places and it
has only one return.  In that scenario, it's very easy for this to happen.

Using INTERNAL should work and that should always be the first row of any
return_info.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>