cse: let equivalent casts hash & compare identically
Now that cast instructions are more finely grained, it's
not needed to compare the original type of the casts,
only the original size can matter.

So, do not hash & compare the original types but only the
orignal sizes. This allow much more casts instructions to be
CSEed away.

Note: like noted in the code, even the original size shouldn't
      matter as identical sources should implies identical
      original sizes but this can't yet be guaranted.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>