add a function to remove deadborn instructions
The values produced by expressions are not always used,
it's normal. However this produce useless code that will
need to be removed, sooner or later.

Currently, this removal of dead instructions is done as
part of the normal instruction simplification where each
kind of instruction are checked if it is used or not and,
if not, the usage is removed from the operands and the
instruction itself is removed. This has several drawbacks:
* this is done at each simplification cycle while it is
  only needed just after linearization
* there is a large overlap between what is needed to correctly
  remove these dead instructions and kill_instruction()
* instructions which are not simplified are forgotten
  for this and thus never removed.

Change this by adding a separate function to remove these
deadborn instructions.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>