function_hooks: change order of how db verse implied returns are handled
The code here is sort of legacy stuff where we have three different ways
of determining the return value.  The issue with this patch concerns the
way strlen() is handled.  We set that up in smatch_common_functions.c where
we have some custom code to say what the return value is.  That code should
over write what the DB says so it should happen after we do the stuff with
the db.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>