db/fixup_kernel: handle __builtin_bswap64()
This is the exact same thing as __fsab64() so it should be OK.  The thing
here is that Smatch ignores byte swaps and instead stores the cpu value
always.  That let's us still have the value if we swap back.

I think it causes one false positive which is really subtle.  Or possibly
that was a manual swap and I got confused.  Anyway, we can't do math with
non-cpu endian data so it's not horrible to store the wrong value there.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>