user_data2: don't set the function itself to user_data
This doesn't affect real life, but I was looking at the debug output and
noticed this mistake.  Say the user can control the return value of a
function, but we don't care because we don't store the return, in that
case, program was setting the function itself to user_data.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>