real_absolute: ignore iterator pre statements
The problem here is that we handle for loops in a short cut way instead of
doing it properly.  This is perhaps not the correct solution, but it
works 99% of the time.

The issue is that we're duplicating what smatch_extra does but we're
doing it badly, so we store "i = 0" in real_absolute, but we're storing
"i = 0-9" in smatch_extra.  In this situation, smatch_extra information
is more accurate than real_absolute.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>