math: fix handling of sizeof() __le16 types
I'm pretty sure the handle_sizeof() code was copy and pasted from Sparse.

There is a comment that implies it is needed but, in testing, it just
seems to create bugs.  I have ifdefed it out and haven't noticed anything
bad.  Anyway, let's run with it for a bit then delete the code.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>