pre-process: try find any include that looks like it might work
When we're parsing code from other arches, then normally we error out
because we can't find the right include files.  With this code, we just
look for any include with that name and include it.  Also instead of dying
we just keep going.

It prints a lot of garbage output mostly, but some of the warnings still
work.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>