type: fix type_bits() for arrays
Say you have an array:

	char foo[100];

And you want to find how many bits are in the type of &foo, it should be
the same as a pointer.  But actually what happens is that it gives you
a completely different number.

This is a problem because we have functions that return specific things
if a parameter is 4095-s64max (as opposed to NULL).  But then smatch gets
confused about the possible values of "foo" and thinks it is "(-128)-127"
so it says all those returns are impossible.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>